pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hlo-exporter"
version = "0.1.0"
description = "A tool for exporting High-Level Operations (HLO) from JAX/ML models"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">=3.11"
keywords = ["jax", "hlo", "ml", "export", "compiler"] # for searching
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
    "Topic :: Scientific/Engineering",
]

dependencies = [
    "aqtp>=0.9.0",
    "flax==0.11.*",
    "jax>=0.7.0",
    "omegaconf>=2.3.0",
    "pytest>=8.4.2",
    "pytest-cov>=5.0.0",
]

[project.optional-dependencies]
# 用于代码风格检查和格式化
lint = [
    "ruff>=0.13.2",
]
# 用于静态类型检查
typing = [
    "mypy>=1.18.2",
]
# dev 组包含了上面所有的开发工具，方便一键安装
dev = [
    "hlo-exporter[lint,typing]",
]

[tool.pytest.ini_options]
pythonpath = [
    ".",
]
addopts = [
    "--cov=.",
    "--cov-report=term-missing",
    "--tb=short",
]

[tool.coverage.run]
source = [
    "hlo_exporter",
    "layers",
    "utils",
    "models",
    "entrypoints",
]
omit = [
    "tests/*",
]

[tool.hatch.build.targets.wheel]
packages = ["layers", "models", "utils"]
include = ["*.py"]